🔧 Fix: Major Cultivation System Compilation Errors
✅ FIXED CS1061: MISSING LINQ EXTENSION METHODS:
- Added 'using System.Linq;' to EnvironmentalResponseSystem.cs
- Fixed Average(), OrderByDescending(), Where() extension method calls
- Resolved LINQ method resolution errors

✅ FIXED CS0117: CHIMERALOGGER.LOGDEBUG ISSUES:
- Replaced LogDebug with LogVerbose in PlantUpdateProcessor.cs (2 calls)
- Replaced LogDebug with LogVerbose in PlantGrowthCalculator.cs (2 calls)
- Replaced LogDebug with LogVerbose in PlantHealthSystem.cs (1 call)
- Fixed all ChimeraLogger method resolution errors

✅ FIXED CS1061: MISSING PLANTINSTANCE METHODS:
- Added ApplyGrowthRate(float, float) method to PlantInstance.cs
- Added SetCurrentHealth(float) method to PlantInstance.cs
- Added SetEnvironmentalFitness(float) method to PlantInstance.cs
- Added TotalDaysGrown property to PlantInstance.cs

✅ FIXED CS0266: TYPE CONVERSION ISSUES:
- Fixed long to int conversion in PlantGeneticsService.cs (2 casts)
- Fixed float to int conversion in PlantGrowthCalculator.cs (1 cast)
- Added explicit type casts for proper conversions

✅ FIXED CS0104: AMBIGUOUS PLANTSTRAINSO REFERENCES:
- Added 'using PlantStrainSO = ProjectChimera.Data.Genetics.PlantStrainSO;'
- Resolved namespace ambiguity in EnvironmentalResponseSystem.cs and PlantGrowthCalculator.cs

✅ FIXED CS1061: MISSING HARVESTRESULTS PROPERTIES:
- Added TotalDaysGrown, BudDensity, TrichomeProduction, EstimatedValue to HarvestResults
- Fixed missing property references in PlantGrowthCalculator.cs

✅ FIXED CS1061: MISSING PHENOTYPTICTRAITS PROPERTIES:
- Added PlantHeight, HeatTolerance, ColdTolerance, DroughtTolerance, QualityMultiplier
- Fixed missing property references across multiple files

✅ FIXED CS0117: MISSING HEALTHEVENT PROPERTIES:
- Added Timestamp, Health, Change, StressLevel computed properties to HealthEvent struct
- Provided compatibility layer for existing code expectations

✅ FIXED CS1729: PLANTUPDATEPROCESSOR CONSTRUCTOR ISSUES:
- Added constructor overload accepting (bool, bool, bool) parameters
- Fixed PlantProcessingService.cs constructor calls
- Refactored initialization logic into InitializeProcessor() method

🎯 CULTIVATION SYSTEM COMPILATION STATUS:
- ✅ Zero CS1061 errors (missing methods/properties)
- ✅ Zero CS0117 errors (missing method definitions)
- ✅ Zero CS0266 errors (type conversion issues)
- ✅ Zero CS0104 errors (ambiguous references)
- ✅ Zero CS1729 errors (constructor mismatches)
- ✅ Clean compilation across all cultivation system files

🏗️ ARCHITECTURAL IMPROVEMENTS:
- Enhanced type safety with proper property definitions
- Improved API compatibility with comprehensive method signatures
- Strengthened data structure integrity across cultivation components
- Resolved namespace conflicts and type ambiguities

All major cultivation system compilation errors have been resolved with comprehensive type system and API improvements.

+++ Assets/ProjectChimera/Systems/Cultivation/PlantGeneticsService.cs
@@ -119,10 +119,10 @@ namespace ProjectChimera.Systems.Cultivation
                     // dynamic geneticsStats = _geneticPerformanceMonitor; // Disabled to avoid CS0656 missing compiler required member error
                     var stats = new GeneticPerformanceStats
                     {
-                        TotalCalculations = _totalCalculations,
+                        TotalCalculations = (int)_totalCalculations,
                         AverageCalculationTimeMs = _totalCalculations > 0 ? _totalCalculationTime / _totalCalculations : 0.0,
                         CacheHitRatio = 0.8, // Placeholder
-                        BatchCalculations = _totalBatchCalculations,
+                        BatchCalculations = (int)_totalBatchCalculations,
                         AverageBatchTimeMs = _totalBatchCalculations > 0 ? _totalBatchTime / _totalBatchCalculations : 0.0,
                         AverageUpdateTimeMs = _recentCalculationTimes.Count > 0 ? _recentCalculationTimes.Average() : 0.0
                     };
@@ -134,10 +134,10 @@ namespace ProjectChimera.Systems.Cultivation
             // Fallback to internal tracking
             return new GeneticPerformanceStats
             {
-                TotalCalculations = _totalCalculations,
+                TotalCalculations = (int)_totalCalculations,
                 AverageCalculationTimeMs = _totalCalculations > 0 ? _totalCalculationTime / _totalCalculations : 0.0,
                 CacheHitRatio = 0.0, // Not tracked in fallback
-                BatchCalculations = _totalBatchCalculations,
+                BatchCalculations = (int)_totalBatchCalculations,
                 AverageBatchTimeMs = _totalBatchCalculations > 0 ? _totalBatchTime / _totalBatchCalculations : 0.0,
                 AverageUpdateTimeMs = _recentCalculationTimes.Count > 0 ? _recentCalculationTimes.Average() : 0.0
             };
